First Commit (Bot.py, Main.py) uploaded
This bot is a simple Discord bot that allows users to check the current price of various cryptocurrencies. The bot can be summoned by sending a message with the command prefix "!" followed by the command "price". The bot will then send an embed message containing buttons for various cryptocurrencies, such as Bitcoin, Ethereum, and Litecoin. Users can select the cryptocurrency they want to check the price of by clicking on the corresponding button. The bot will then send a message containing the current price of the selected cryptocurrency in USD.

This bot uses the CoinGecko API to retrieve the current prices of the cryptocurrencies. The API returns the prices in JSON format, which is then parsed and used to construct the messages sent by the bot. The bot uses the Discord.py library to interact with the Discord API and handle the various events.

This bot is a simple and easy way to check the prices of various cryptocurrencies without having to leave the Discord application. It can be useful for people who are interested in cryptocurrency trading or just want to keep an eye on the prices of their favorite coins.